Update index.html
The only change should be that the css should be in a separate file, and linked so that it still affects the page. I think this needs to be implemented to allow me to see if it worked completely. Please let me know any errors on my end that need to be fixed.