Update index.html (#15)
* Update index.html

Made the tabs cleaner (spacing and color, etc), added a title and slogan in the proper place. Couldn't get the image to move from where it currently is. I can't figure out how to make <link> tag work since I can't seem to get the file path for the main.css file in our github. When I used <link src = "main.css"> it didn't work. If you know how to get the correct file path/name, that would be great :)

* Update index.html

The only change should be that the css should be in a separate file, and linked so that it still affects the page. I think this needs to be implemented to allow me to see if it worked completely. Please let me know any errors on my end that need to be fixed.

* Update index.html

removed strong tags and renamed the openCity function

* Update index.html

Co-authored-by: CriticalWasTaken <104160959+CriticalWasTaken@users.noreply.github.com>